+++ test/test_interface.py
@@ -112,8 +112,8 @@ def test_check_kernel_correctness(dev_func_interface):
     wrong = [numpy.array([1,2,3,4]).astype(numpy.float32)]
     atol = 1e-6
 
-    instance = core.KernelInstance("name", "kernel_string", "temp_files", "threads", "grid", {})
-    test = dev.check_kernel_correctness('func', answer, instance, answer, atol, True)
+    instance = core.KernelInstance("name", "kernel_string", "temp_files", "threads", "grid", {}, answer)
+    test = dev.check_kernel_correctness('func', answer, instance, answer, atol, None, True)
 
     dfi.memset.assert_called_once_with(answer[0], 0, answer[0].nbytes)
     dfi.run_kernel.assert_called_once_with('func', answer, 'threads', 'grid')
@@ -133,7 +133,7 @@ def test_check_kernel_correctness(dev_func_interface):
     #obviously does not result in the result_host array containing anything
     #non-zero
     try:
-        dev.check_kernel_correctness('func', wrong, instance, wrong, atol, True)
+        dev.check_kernel_correctness('func', wrong, instance, wrong, atol, None, True)
         print("check_kernel_correctness failed to throw an exception")
         assert False
     except Exception: